Rename and group styles for correlation tiles
